+++ src/page/award_infor/index.js
@@ -4,29 +4,14 @@ import renderRedHtml from '../find_award/template/redHtmlTemplate';
 import renderGreenHtml from '../find_award/template/greenHtmlTemplate';
 // localStorage.setItem('activityId', 1);
 import { delegate } from '~/util/elemnet';
-function GetRequest() {
-  let url = location.search; // 获取url中"?"符后的字串
-  let theRequest = new Object();
-  if (url.indexOf('?') != -1) {
-    let str = url.substr(1);
-    strs = str.split('&');
-    for (let i = 0; i < strs.length; i++) {
-      theRequest[strs[i].split('=')[0]] = (strs[i].split('=')[1]);
-    }
-  }
-  return theRequest;
-}
-let Request = new Object();
-Request = GetRequest();
-name = Request['name'];
 const id = localStorage.getItem('activityId');
+console.log(id);
 document.getElementById('redeem').addEventListener('click', redeem, false);
 document.getElementById('notredeem').addEventListener('click', notredeem, false);
 notredeem()
 function notredeem() {
     let div = document.getElementById('display');
     displayNotRedeem(id).then((data) => {
-        
         let persion = data.data;
         let htmltext = '';
         for (let i = 0; i < persion.length; i++) {
@@ -37,8 +22,7 @@ function notredeem() {
     
     })
     const attachEvent = document.getElementById('display');
-    delegate(attachEvent, '#inerEvent', 'click', duijiang, false);
-    
+    delegate(attachEvent, '#inerEvent', 'click', duijiang, false);   
 }
 
 function redeem() {
