+++ src/page/lottery/index.js
@@ -2,7 +2,23 @@ import { displayActivity } from '~/ajax/activity';
 import codes from '~/config/codeConfig';
 import { drawPrize } from '~/ajax/lottery';
 // 活动数据渲染
-const a = displayActivity(9).then((data) => {
+function GetRequest() {
+  let url = location.search; // 获取url中"?"符后的字串
+  console.log(url);
+  let theRequest = new Object();
+  if (url.indexOf('?') != -1) {
+    let str = url.substr(1);
+    let strs = str.split('&');
+    for (let i = 0; i < strs.length; i++) {
+      theRequest[strs[i].split('=')[0]] = (strs[i].split('=')[1]);
+    }
+  }
+  return theRequest;
+}
+let Request = new Object();
+Request = GetRequest();
+let activityId = Request['activityId'];
+const a = displayActivity(activityId).then((data) => {
   let _data;
   if (data.code == codes.success) {
    _data = data.data.awards;
